backend/requirements.txt
uvicorn==0.29.0
transformers==4.40.0
torch==2.1.2
moviepy==1.0.3
soundfile==0.12.1
yt-dlp==2024.05.27
python-multipart==0.0.9
fastapi[all]==0.110.0
python-jose[cryptography]